simplify auth
